Add a new API endpoint `GET /v2/users/sparseMatches?qs=...`
Added new API endpoint `GET /v2/users/sparseMatches?qs=...` to get list of non-private users and groups, whose username _sparse matches_ the query string. The sparse match means that the username must contain all the query string letters in the given order, but with possible other letters in between. For example, the 'ur[an]us', '[an]tenna', 's[a]tur[n]' matches the "an" query.

The query string should match the `[a-z0-9-]{2,}` regex.

The order of the results is not specified, the client should sort them at  their side.